Refactored custom schema parser to a class
